third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-z-index-inherit-expected.txt

This is a testharness.js-based test.
FAIL z-index should inherit the computed value assert_equals: expected "1" but got "0"
Harness: the test ran to completion.
